+++ packages/x-components/src/types/page-mode.ts
@@ -1,0 +1,2 @@
+// Indicates whether the results will be loaded using infinite scroll or a page selector.
+export type PageMode = 'infinite_scroll' | 'pagination';