+++ packages/x-adapter-platform/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.0-alpha.10](https://github.com/empathyco/x/compare/@empathyco/x-adapter-platform@1.1.0-alpha.9...@empathyco/x-adapter-platform@1.1.0-alpha.10) (2025-01-22)
+
+**Note:** Version bump only for package @empathyco/x-adapter-platform
+
+
+
+
+
 ## [1.1.0-alpha.9](https://github.com/empathyco/x/compare/@empathyco/x-adapter-platform@1.1.0-alpha.8...@empathyco/x-adapter-platform@1.1.0-alpha.9) (2024-11-25)
 
 **Note:** Version bump only for package @empathyco/x-adapter-platform